split file from base.rb
